Fixed compilation errors on gcc4.
git-svn-id: http://svn.boost.org/svn/boost/trunk@29028 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/wave/samples/cpp_tokens/slex/cpp_slex_lexer.hpp
@@ -526,7 +526,9 @@ public:
             string_type token_val(value.c_str());
             
                 if (T_CONTLINE != id) {
-                    switch (id) {
+                //  The cast should avoid spurious warning about missing case labels 
+                //  for the other token ids's.
+                    switch ((unsigned int)id) {   
                     case T_IDENTIFIER:
                     // test identifier characters for validity (throws if 
                     // invalid chars found)
